pyproject.toml
[tool.poetry]
name = "ocrhelper"
version = "0.3.0"
description = "Tool for translating text from the screen and some more"
authors = ["Sergei <reyangood@gmail.com>"]
readme = "README.md"
packages = [{include = "ocrhelper"}]

[tool.poetry.dependencies]
python = "^3.10"
pyautogui = "^0.9.54"
pillow = "9.5.0"
numpy = "^1.25.1"
deep-translator = {extras = ["ai"], version = "^1.11.4"}

[tool.poetry.group.dev.dependencies]
pytest = "^7.4.0"
flake8 = "^6.0.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
